requirements.txt
# Microsoft Graph API dependencies
azure-identity>=1.12.0
msgraph-sdk>=1.0.0

# Google Auth & GCP SDK dependencies
google-auth>=2.16.0
google-api-python-client>=2.70.0
google-cloud-storage>=2.8.0
google-cloud-bigquery>=3.4.0

# Airflow (Composer) dependencies
apache-airflow>=2.5.0
google-cloud-orchestration-airflow>=1.4.0
apache-airflow-providers-google>=8.10.0
apache-airflow-providers-microsoft-azure>=5.2.0

# Common dependencies
requests>=2.28.0
python-dotenv>=1.0.0
